+++ packages/blockingpy-core/blockingpy/text_encoders/base.py
@@ -10,7 +10,6 @@ from ..data_handler import DataHandler
 
 
 class TextEncoder(ABC):
-
     """
     Abstract base class for text-to-matrix transformers.